parser/requirements.txt
pypdf>=4.0.0
pdfplumber>=0.10.0
pdfminer.six>=20221105